Fix shared.table.Draw()
There was a bug that resulted in a fatal crash when printing datatypes that were not strings using var.(string). This has been fixed by using fmt.Sprint(). Additionally, an index out of range error has been fixed that was due to the indices in a nested for loop being switched in line 65.